requirements.txt
alembic==1.4.0
Babel==2.8.0
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
python-dateutil==2.6.0
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
WTForms==2.2.1